pdf-creator-server/requirements.txt
fastmcp
playwright
python-dotenv
Pillow
